add endpoint to recalculate truedl
