+++ config/application.rb
@@ -9,6 +9,7 @@ require 'active_job/railtie'
 require 'active_record/railtie'
 require 'action_controller/railtie'
 require 'action_view/railtie'
+require 'true_dl'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
