+++ config/routes.rb
@@ -5,6 +5,7 @@ Rails.application.routes.draw do
   get 'ping', to: 'healthcheck#ping'
   get 'reset_cache', to: 'cache#reset'
   get 'recreate_views/:model', to: 'materialized_views#recreate_views'
+  get 'recalculate_truedl/:model', to: 'true_dls#recalculate'
   get 'mau', to: 'reports#mau'
 
   get ':model/tournaments/', to: 'tournaments#index', as: 'tournaments'
