+++ lib/tasks/truedl.rake
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative '../../app/lib/truedl'
+require_relative '../../app/lib/truedl_calculator'
 
 namespace :true_dl do
   task :calculate_for_tournament, %i[id model] => :environment do |_t, args|
-    TrueDL.calculate_for_tournament(tournament_id: args[:id], model_name: args[:model])
+    TrueDLCalculator.calculate_for_tournament(tournament_id: args[:id], model_name: args[:model])
   end
 
   task :calculate_for_all_maii_tournaments, [:model] => :environment do |_t, args|
-    TrueDL.calculate_for_all_maii_tournaments(model_name: args[:model])
+    TrueDLCalculator.calculate_for_all_maii_tournaments(model_name: args[:model])
   end
 end
